requirements.txt
streamlit
pandas
google-generativeai
plotly
streamlit-option-menu
firebase-admin
pillow
streamlit-paste-button